requirements.txt
chromadb
sentence-transformers
google-generativeai
pandas
numpy
python-dotenv
torch==2.5.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
